requirements/local.txt
alabaster==0.7.11
amqp==2.5.2
attrdict==2.0.1
asgiref==3.2.7
argon2-cffi==18.1.0
atomicwrites==1.1.5
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
billiard==3.6.1.0
boto3==1.8.5
base58==2.0.0
celery==4.4.0
celery-once==3.0.1
flower==0.9.3
certifi==2019.11.28
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
cytoolz==0.10.1
decorator==4.3.0
defusedxml==0.5.0
Django==2.2.9
django-coverage-plugin==1.5.0
django-crispy-forms==1.7.2
django-environ==0.4.5
django-extensions==2.1.0
django-model-utils==3.1.2
django-redis==4.11.0
djangorestframework==3.9.4
docutils==0.14
eth-abi==2.1.0
eth-account==0.4.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.1.2
eth-typing==2.2.1
eth-utils==1.8.4
factory-boy==2.11.1
Faker==0.8.17
flake8==3.5.0
gunicorn==19.9.0
hexbytes==0.2.0
idna==2.8
imagesize==1.0.0
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
ipfshttpclient==0.4.12
importlib-metadata==1.4.0
itypes==1.1.0
jedi==0.12.1
Jinja2==2.10
jsonschema==3.2.0
kombu==4.6.7
lru-dict==1.1.6
mccabe==0.6.1
more-itertools==8.1.0
numpy==1.16.3
multiaddr==0.0.9
mypy-extensions==0.4.3
netaddr==0.7.19
parsimonious==0.8.1
protobuf==3.11.2
py-solc==3.2.0
rlp==1.2.0
oauthlib==2.1.0
packaging==17.1
parso==0.3.1
pickleshare==0.7.4
Pillow==5.4.1
pluggy==0.6.0
prompt-toolkit==1.0.15
psycopg2-binary==2.7.5
py==1.5.4
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
Pygments==2.2.0
pyparsing==2.2.0
pytest==3.6.3
pytest-django==3.3.2
pytest-sugar==0.9.1
python-dateutil==2.7.3
python-slugify==1.2.5
pytz==2019.3
redis==3.3.11
requests==2.22.0
simplegeneric==0.8.1
six==1.14.0
semantic-version==2.8.4
snowballstemmer==1.2.1
Sphinx==1.7.6
sphinxcontrib-websupport==1.1.0
sqlparse==0.3.0
termcolor==1.1.0
toolz==0.10.0
tornado==5.1.1
typing-extensions==3.7.4.1
text-unidecode==1.2
traitlets==4.3.2
ulid-py==0.0.6
Unidecode==1.0.22
uritemplate==3.0.0
urllib3==1.23
varint==1.0.2
vine==1.3.0
wcwidth==0.1.7
Werkzeug==0.14.1
web3==5.4.0
websockets==8.1
zipp==2.0.0

# Django extensions
rest_condition==1.0.3
django-memoize==2.1.1
django-storages==1.6.6  # https://github.com/jschneier/django-storages
django-ses>=0.8.6  # https://github.com/django-ses/django-ses
redis-py-cluster==1.3.5  # for redis cluster

# client library
elasticsearch==6.3.1

# Crypto
pycryptodome==3.8.2
PyDes==2.0.1

# Channels
# channels==2.0.2
# channels_redis==2.1.1

# local only
django-queryinspect==1.0.0

tornado==5.1.1

# Excel
xlwt==1.3.0

# qrcode
qrcode==6.1

django-db-logger==0.1.7

reportlab==3.5.42

line-bot-sdk==1.16.0
pyzbar==0.1.8
opencv-python==4.4.0.42
pyOpenSSL==19.1.0
